USB HID initialise fail try loop
When the host device is not powered or goes through a power cycle e.g. when connected to a portable computer or through a KVM. The USB HID detection may raise an error and stop the code.

A try and repeat loop is implemented to loop and flash the LED on every error iteration until the error clears i.e. the macro pad device is turned off or it is plugged into a host device.